feat(cli): add Add command wrapper and refactor business logic
Created internal/cli/add.go as Cobra wrapper for add command.
Moved business logic from cmd/*_logic.go to business/ package
to resolve circular dependency between cmd and internal/cli packages.

Changes:
- Created internal/cli/add.go: Cobra command wrapper that calls
  business.Add() and formats output to match todo.txt-cli
- Moved cmd/*_logic.go -> business/*_logic.go: Business logic functions
- Removed old Cobra wrappers from cmd/ (add.go, list.go, do.go, pri.go,
  tidy.go, config.go)
- Updated cmd/root.go: Commented out command registration (temporary)

Architectural improvement: Separating business logic into its own
package (business/) provides cleaner separation of concerns and
prevents circular dependencies between CLI layer and command logic.
