+++ business/add_logic.go
@@ -1,4 +1,4 @@
-package cmd
+package business
 
 import (
 	"fmt"
