+++ business/do_logic_test.go
@@ -1,4 +1,4 @@
-package cmd
+package business
 
 import (
 	"bytes"
@@ -7,8 +7,8 @@ import (
 	"github.com/gkarolyi/togodo/todotxtlib"
 )
 
-func TestSetPriority(t *testing.T) {
-	t.Run("sets priority on task", func(t *testing.T) {
+func TestDo(t *testing.T) {
+	t.Run("toggles task done", func(t *testing.T) {
 		// Setup
 		var buf bytes.Buffer
 		buf.WriteString("task one\n")
@@ -17,17 +17,17 @@ func TestSetPriority(t *testing.T) {
 		repo, _ := todotxtlib.NewFileRepository(reader, writer)
 
 		// Execute
-		result, err := SetPriority(repo, []int{0}, "A")
+		result, err := Do(repo, []int{0})
 
 		// Assert
 		if err != nil {
 			t.Fatalf("unexpected error: %v", err)
 		}
-		if len(result.UpdatedTodos) != 1 {
-			t.Errorf("expected 1 updated todo, got %d", len(result.UpdatedTodos))
+		if len(result.ToggledTodos) != 1 {
+			t.Errorf("expected 1 toggled todo, got %d", len(result.ToggledTodos))
 		}
-		if result.UpdatedTodos[0].Priority != "A" {
-			t.Errorf("expected priority A, got %s", result.UpdatedTodos[0].Priority)
+		if !result.ToggledTodos[0].Done {
+			t.Error("expected todo to be marked done")
 		}
 	})
 
@@ -40,7 +40,7 @@ func TestSetPriority(t *testing.T) {
 		repo, _ := todotxtlib.NewFileRepository(reader, writer)
 
 		// Execute with invalid index
-		_, err := SetPriority(repo, []int{99}, "A")
+		_, err := Do(repo, []int{99})
 
 		// Assert
 		if err == nil {
@@ -48,7 +48,7 @@ func TestSetPriority(t *testing.T) {
 		}
 	})
 
-	t.Run("validates all indices before setting priority on any", func(t *testing.T) {
+	t.Run("validates all indices before toggling any", func(t *testing.T) {
 		// Setup
 		var buf bytes.Buffer
 		buf.WriteString("task one\ntask two\n")
@@ -57,17 +57,17 @@ func TestSetPriority(t *testing.T) {
 		repo, _ := todotxtlib.NewFileRepository(reader, writer)
 
 		// Execute with one valid and one invalid index
-		_, err := SetPriority(repo, []int{0, 99}, "A")
+		_, err := Do(repo, []int{0, 99})
 
 		// Assert error occurred
 		if err == nil {
 			t.Fatal("expected error for invalid index, got nil")
 		}
 
-		// Verify task 0 priority was NOT set (atomicity preserved)
+		// Verify task 0 was NOT toggled (atomicity preserved)
 		todos, _ := repo.ListAll()
-		if todos[0].Priority != "" {
-			t.Errorf("task 0 priority should not be set when operation fails, got %s", todos[0].Priority)
+		if todos[0].Done {
+			t.Error("task 0 should not be toggled when operation fails")
 		}
 	})
 }
