+++ business/list_logic_test.go
@@ -1,4 +1,4 @@
-package cmd
+package business
 
 import (
 	"bytes"
